Fix an infinite recursion where the wrong overload was called

The intension was that if Camera::Zoom was called with a Float64, it
would call the Float32 version. However, I made a type where it called
itself. The latest version of XCode called me out on this infinite
recursion.

I think this is the equivalent of the compiler calling me a dumb-ass.

See merge request !483

+++ vtkm/rendering/Camera.h
@@ -537,7 +537,7 @@ public:
   VTKM_CONT_EXPORT
   void Zoom(vtkm::Float64 zoom)
   {
-    this->Zoom(static_cast<vtkm::Float64>(zoom));
+    this->Zoom(static_cast<vtkm::Float32>(zoom));
   }
 
   /// \brief Moves the camera as if a point was dragged along a sphere.
